feat: Add Qt 6 support (#209)
* qt6: webengine(quick) module no longer required

Relevant componets were moved to webenginecore

* qt6: QWebEngineDownloadRequest

QWebEngineDownloadItem was renamed to QWebEngineDownloadRequest and the downloadProgress signal was split into receivedBytesChanged and totalBytesChanged signals

* qt6: QWebEnginePage::view() polyfill

QWebEnginePage was moved to the webenginecore module in Qt6. As such, it cannot reference QWidget from the widgets module. QWebEnginePage::view() was replaced by QWebEngineView::forPage() from the webenginewidgets module.

* qt6: QWebEnginePage::certificateError() is now a signal

QWebEnginePage::certificateError() was changed from an overridable method to a signal in Qt6. The response actions are now methods on the QWebEngineCertificateError object. For some reason, said object is passed as const reference and must be const_cast before applying a resolution.

* qt6: QWebEngineContextMenuRequest

QWebEngineContextMenuData was renamed to QWebEngineContextMenuRequest in Qt6. The related methods were moved from QWebEnginePage (which moved to the webenginecore module) to QWebEngineView.

* qt6: QWebEngineSettings::defaultSettings() removal

QWebEngineSettings::defaultSettings() was removed in Qt6, use QWebEngineProfile::defaultProfile()->settings() insteadd.

* qt6: QStandardPaths::DataLocation removal

QStandardPaths::DataLocation was already deprecated in Qt5 and removed in Qt6. It was an alias for QStandardPaths::AppLocalDataLocation.

* qt6: QDesktopWidget removal

QDesktopWidget (and QApplication::desktop()) was removed in Qt6

* qt6: Use QString::fromUtf8() instead of QTextCodec

QTextCodec was removed (movec to core5compat) in Qt6.

* qt6: QGeoPositionInfoSource::updateTimeout removal

The QGeoPositionInfoSource::updateTimeout signal was removed in Qt6, with it's functionality taken over by the QGeoPositionInfoSource::errorOccured signal.

* qt6: QDateTime::toTime_t() removal

QDateTime::toTime_t() was already deprecated and replaced with QDateTime::toSecsSinceEpoch() in Qt5. It was removed in Qt6.

* qt6: Fix some deprecation warnings